feat: merge flask view
- Merge the flask admin view "associer un role à l'export inside
  "Export" view
